Refactor sync out into a submodule
